#365 restructuring of components
