+++ client/src/components/form-groups/SymptomsForm.vue
