+++ client/src/components/inputs/BarcodeScanner.vue
