+++ client/src/components/inputs/LaboratoryInput.vue
