+++ client/src/components/other/IndexPatientTableCell.vue
