+++ client/src/views/Account.vue
@@ -172,9 +172,9 @@
 import Vue from 'vue'
 import { InstitutionType, UserRole } from '@/models'
 import { authMapper } from '@/store/modules/auth.module'
-import ChangePasswordForm from '@/components/ChangePasswordForm.vue'
-import ChangeInstitutionForm from '@/components/ChangeInstitutionForm.vue'
-import AddOrChangeUserForm from '@/components/AddOrChangeUserForm.vue'
+import ChangePasswordForm from '@/components/modals/ChangePasswordForm.vue'
+import ChangeInstitutionForm from '@/components/modals/ChangeInstitutionForm.vue'
+import AddOrChangeUserForm from '@/components/modals/AddOrChangeUserForm.vue'
 import { UserDTO } from '@/api/SwaggerApi'
 
 const tableColumns = [
