+++ client/src/views/PatientDetails.vue
@@ -560,11 +560,11 @@ import { SYMPTOMS } from '@/models/symptoms'
 import { PRE_ILLNESSES } from '@/models/pre-illnesses'
 import { Column } from 'ant-design-vue/types/table/column'
 import { TestTypeItem, testTypes } from '@/models/test-types'
-import ChangePatientStammdatenForm from '@/components/ChangePatientStammdatenForm.vue'
-import EditExposureContact from '@/components/EditExposureContact.vue'
+import ChangePatientStammdatenForm from '@/components/modals/ChangePatientStammdatenForm.vue'
+import EditExposureContact from '@/components/form-groups/EditExposureContact.vue'
 import { map } from '@/util/mapping'
 import { Modal } from 'ant-design-vue'
-import ChangePatientFalldatenForm from '@/components/ChangePatientFalldatenForm.vue'
+import ChangePatientFalldatenForm from '@/components/modals/ChangePatientFalldatenForm.vue'
 import { EXPOSURE_LOCATIONS, EXPOSURES_INTERNAL } from '@/models/exposures'
 
 const columnsTests: Partial<Column>[] = [
