+++ client/src/views/SubmitTestResult.vue
@@ -120,9 +120,9 @@
 import { Institution, LabTest } from '@/api/SwaggerApi'
 import Vue from 'vue'
 import Api from '@/api'
-import TestInput from '@/components/TestInput.vue'
-import LaboratoryInput from '@/components/LaboratoryInput.vue'
-import DateInput from '@/components/DateInput.vue'
+import TestInput from '@/components/inputs/TestInput.vue'
+import LaboratoryInput from '@/components/inputs/LaboratoryInput.vue'
+import DateInput from '@/components/inputs/DateInput.vue'
 import { authMapper } from '@/store/modules/auth.module'
 import { testResults, TestResultType } from '@/models/event-types'
 import moment from 'moment'
